refactor: BaseComponentとthemeをsharedレイヤーに移行（TASK-0067）
- src/shared/components/BaseComponent.tsを作成
- src/shared/theme/theme.tsを作成
- 既存のpresentation層のファイルは後方互換性のため再エクスポートに変更
- @deprecatedタグで移行先を案内
