+++ atelier-guild-rank/src/shared/components/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Shared Components Module
+ * 共通コンポーネントの公開エクスポート
+ *
+ * TASK-0067: shared/components作成
+ */
+
+export type { BaseComponentOptions } from './BaseComponent';
+export { BaseComponent } from './BaseComponent';
