+++ atelier-guild-rank/src/shared/index.ts
@@ -3,7 +3,9 @@
  * 共有コードの公開エクスポート
  */
 
+export * from './components';
 export * from './constants';
 export * from './services';
+export * from './theme';
 export * from './types';
 export * from './utils';
