+++ atelier-guild-rank/src/shared/theme/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Shared Theme Module
+ * テーマ定義の公開エクスポート
+ *
+ * TASK-0067: shared/theme作成
+ */
+
+export type {
+  BackgroundColorKey,
+  BorderColorKey,
+  CardTypeColorKey,
+  ColorKey,
+  TextColorKey,
+} from './theme';
+export { Colors, THEME } from './theme';
